Stop returning modified values, get IP and port in process.
I've realized that when we're passed a value of some sort, when we
modify it, it modifies the original, since we're accessing a
pointer (or something like that). Also, we now get the client IP
and port ourself, and pass it back to the caller when we return
a value.